+++ src/main/java/com/cells/blocks/importinterface/TileImportInterface.java
@@ -385,19 +385,19 @@ public class TileImportInterface extends AENetworkInvTile implements IGridTickab
      * Format a tick count as a human-readable time string.
      * Format: "1d 2h 3m 4s" (skipping zero parts) or "0" if zero.
      */
-    public static String formatPollingRate(int ticks) {
+    public static String formatPollingRate(long ticks) {
         if (ticks <= 0) return "0";
 
-        int days = ticks / TICKS_PER_DAY;
+        long days = ticks / TICKS_PER_DAY;
         ticks %= TICKS_PER_DAY;
 
-        int hours = ticks / TICKS_PER_HOUR;
+        long hours = ticks / TICKS_PER_HOUR;
         ticks %= TICKS_PER_HOUR;
 
-        int minutes = ticks / TICKS_PER_MINUTE;
+        long minutes = ticks / TICKS_PER_MINUTE;
         ticks %= TICKS_PER_MINUTE;
 
-        int seconds = ticks / TICKS_PER_SECOND;
+        long seconds = ticks / TICKS_PER_SECOND;
 
         StringBuilder sb = new StringBuilder();
         if (days > 0) sb.append(days).append("d ");
